+++ src/main/java/com/vivero/vivero_backend/api/service/ProductoService.java
@@ -2,6 +2,7 @@ package com.vivero.vivero_backend.api.service;
 
 
 import com.vivero.vivero_backend.api.model.Producto;
+import com.vivero.vivero_backend.api.repository.DetalleVentaRepository;
 import com.vivero.vivero_backend.api.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,9 @@ public class ProductoService {
 
     @Autowired
     private ProductRepository productoRepository;
+    
+    @Autowired
+    private DetalleVentaRepository detalleVentaRepository;
 
     public List<Producto> listarTodos() {
         return productoRepository.findAll();
@@ -48,6 +52,12 @@ public class ProductoService {
     }
 
     public void eliminar(Long id) {
+        if (!productoRepository.existsById(id)) {
+			throw new RuntimeException("Producto no encontrado con ID: " + id);
+		}
+        if (detalleVentaRepository.existsByProductoId(id)) {
+        	throw new RuntimeException("No se puede eliminar el producto con ID: " + id + " porque tiene detalles de venta asociados.");
+        }
         productoRepository.deleteById(id);
     }
 }
